release 2025 r2

+++ core/src/main/java/dk/digitalidentity/os2skoledata/controller/rest/ConfigRestController.java
@@ -1,6 +1,8 @@
 package dk.digitalidentity.os2skoledata.controller.rest;
 
+import dk.digitalidentity.os2skoledata.config.OS2SkoleDataConfiguration;
 import dk.digitalidentity.os2skoledata.dao.model.ClassroomAdmin;
+import dk.digitalidentity.os2skoledata.dao.model.PasswordAdmin;
 import dk.digitalidentity.os2skoledata.dao.model.PasswordSetting;
 import dk.digitalidentity.os2skoledata.dao.model.StudentPasswordChangeConfiguration;
 import dk.digitalidentity.os2skoledata.dao.model.enums.GradeGroup;
@@ -8,6 +10,8 @@ import dk.digitalidentity.os2skoledata.dao.model.enums.RoleSettingType;
 import dk.digitalidentity.os2skoledata.dao.model.enums.StudentPasswordChangerSTILRoles;
 import dk.digitalidentity.os2skoledata.security.RequireAdministratorRole;
 import dk.digitalidentity.os2skoledata.service.ClassroomAdminService;
+import dk.digitalidentity.os2skoledata.service.InstitutionService;
+import dk.digitalidentity.os2skoledata.service.PasswordAdminService;
 import dk.digitalidentity.os2skoledata.service.PasswordSettingService;
 import dk.digitalidentity.os2skoledata.service.StudentPasswordChangeConfigurationService;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -20,7 +24,9 @@ import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.ResponseBody;
 import org.springframework.web.bind.annotation.RestController;
 
+import java.util.ArrayList;
 import java.util.List;
+import java.util.Set;
 
 @RequireAdministratorRole
 @RestController
@@ -35,8 +41,21 @@ public class ConfigRestController {
     @Autowired
     private ClassroomAdminService classroomAdminService;
 
+    @Autowired
+    private PasswordAdminService passwordAdminService;
+
+    @Autowired
+    private OS2SkoleDataConfiguration configuration;
+
+    @Autowired
+    private InstitutionService institutionService;
+
     @DeleteMapping("/rest/config/{id}/delete")
     public ResponseEntity<?> deleteConfig(@PathVariable long id) {
+        if (!configuration.getStudentAdministration().isEnabled()) {
+            return new ResponseEntity<>(HttpStatus.FORBIDDEN);
+        }
+
         studentPasswordChangeConfigurationService.delete(id);
         return ResponseEntity.ok().build();
     }
@@ -45,6 +64,10 @@ public class ConfigRestController {
     @ResponseBody
     @PostMapping("/rest/config/save")
     public ResponseEntity<?> saveConfig(@RequestBody SaveClientDTO dto) {
+        if (!configuration.getStudentAdministration().isEnabled()) {
+            return new ResponseEntity<>(HttpStatus.FORBIDDEN);
+        }
+
         StudentPasswordChangeConfiguration match = studentPasswordChangeConfigurationService.findById(dto.id());
         if (match == null) {
             match = new StudentPasswordChangeConfiguration();
@@ -63,6 +86,10 @@ public class ConfigRestController {
     @ResponseBody
     @PostMapping("/rest/config/passwordsettings/save")
     public ResponseEntity<?> savePasswordSettings(@RequestBody SavePasswordSettingsDTO dto) {
+        if (!configuration.getStudentAdministration().isEnabled()) {
+            return new ResponseEntity<>(HttpStatus.FORBIDDEN);
+        }
+
         PasswordSetting match = passwordSettingService.getSettings(dto.gradeGroup());
         match.setMinLength(dto.minLength());
         match.setRequireComplexPassword(dto.requireComplexPassword());
@@ -73,6 +100,10 @@ public class ConfigRestController {
 
     @DeleteMapping("/rest/config/classroomadmin/{id}/delete")
     public ResponseEntity<?> deleteClassroomAdmin(@PathVariable long id) {
+        if (!configuration.getClassroomAdministration().isEnabled()) {
+            return new ResponseEntity<>(HttpStatus.FORBIDDEN);
+        }
+
         classroomAdminService.delete(id);
         return ResponseEntity.ok().build();
     }
@@ -81,11 +112,49 @@ public class ConfigRestController {
     @ResponseBody
     @PostMapping("/rest/config/classroomadmin/save")
     public ResponseEntity<?> saveClassroomAdmin(@RequestBody ClassroomAdminDTO dto) {
+        if (!configuration.getClassroomAdministration().isEnabled()) {
+            return new ResponseEntity<>(HttpStatus.FORBIDDEN);
+        }
 
         ClassroomAdmin classroomAdmin = new ClassroomAdmin();
         classroomAdmin.setRole(dto.role);
         classroomAdmin.setUsername(dto.username);
         classroomAdminService.save(classroomAdmin);
+
+        return ResponseEntity.ok().build();
+    }
+
+    @DeleteMapping("/rest/config/passwordadmin/{id}/delete")
+    public ResponseEntity<?> deletePasswordadmin(@PathVariable long id) {
+        if (!configuration.getStudentAdministration().isEnabled()) {
+            return new ResponseEntity<>(HttpStatus.FORBIDDEN);
+        }
+
+        passwordAdminService.delete(id);
+        return ResponseEntity.ok().build();
+    }
+
+    private record PasswordAdminDTO(String username, List<Long> institutionIds) {}
+    @ResponseBody
+    @PostMapping("/rest/config/passwordadmin/save")
+    public ResponseEntity<?> savePasswordadmin(@RequestBody PasswordAdminDTO dto) {
+        if (!configuration.getStudentAdministration().isEnabled()) {
+            return new ResponseEntity<>(HttpStatus.FORBIDDEN);
+        }
+
+        if (dto.username == null || dto.institutionIds == null || dto.institutionIds.isEmpty()) {
+            return new ResponseEntity<>("Brugernavn skal udfyldes og institutioner skal vælges", HttpStatus.BAD_REQUEST);
+        }
+
+        if (passwordAdminService.getByUsername(dto.username) != null) {
+            return new ResponseEntity<>("Der findes allerede en IT-kodeordsadministrator med dette brugernavn", HttpStatus.BAD_REQUEST);
+        }
+
+        PasswordAdmin passwordAdmin = new PasswordAdmin();
+        passwordAdmin.setUsername(dto.username);
+        passwordAdmin.setInstitutions(new ArrayList<>());
+        passwordAdmin.getInstitutions().addAll(institutionService.findByIdIn(dto.institutionIds));
+        passwordAdminService.save(passwordAdmin);
 
         return ResponseEntity.ok().build();
     }
